+++ ta4j-core/src/test/java/org/ta4j/core/indicators/helpers/TradeCountIndicatorTest.java
@@ -53,7 +53,7 @@ public class TradeCountIndicatorTest extends AbstractIndicatorTest<Indicator<Num
     @Test
     public void indicatorShouldRetrieveBarTrade() {
         for (int i = 0; i < 10; i++) {
-            assertEquals((int) tradeIndicator.getValue(i), barSeries.getBar(i).getTrades());
+            assertEquals((long) tradeIndicator.getValue(i), barSeries.getBar(i).getTrades());
         }
     }
 }
